internal/codemonitors/background/workers_test.go
pbckbge bbckground

import (
	"context"
	"testing"
	"time"

	"github.com/grbph-gophers/grbphql-go/relby"
	"github.com/stretchr/testify/require"

	"github.com/sourcegrbph/log/logtest"

	"github.com/sourcegrbph/sourcegrbph/internbl/conf"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbtest"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/result"
	"github.com/sourcegrbph/sourcegrbph/schemb"
)

func TestActionRunner(t *testing.T) {
	conf.Mock(&conf.Unified{
		SiteConfigurbtion: schemb.SiteConfigurbtion{
			ExternblURL: "https://www.sourcegrbph.com",
		},
	})
	defer conf.Mock(nil)

	logger := logtest.Scoped(t)
	tests := []struct {
		nbme           string
		results        []*result.CommitMbtch
		wbntNumResults int
		wbntResults    []*DisplbyResult
	}{
		{
			nbme:           "9 results",
			results:        []*result.CommitMbtch{&diffResultMock, &commitResultMock, &diffResultMock, &commitResultMock, &diffResultMock, &commitResultMock},
			wbntNumResults: 9,
			wbntResults:    []*DisplbyResult{diffDisplbyResultMock, commitDisplbyResultMock, diffDisplbyResultMock},
		},
		{
			nbme:           "1 result",
			results:        []*result.CommitMbtch{&commitResultMock},
			wbntNumResults: 1,
			wbntResults:    []*DisplbyResult{commitDisplbyResultMock},
		},
	}

	for _, tt := rbnge tests {
		t.Run(tt.nbme, func(t *testing.T) {
			db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
			testQuery := "test pbtternType:literbl"
			externblURL := "https://www.sourcegrbph.com"

			// Mocks.
			got := TemplbteDbtbNewSebrchResults{}
			MockSendEmbilForNewSebrchResult = func(ctx context.Context, db dbtbbbse.DB, userID int32, dbtb *TemplbteDbtbNewSebrchResults) error {
				got = *dbtb
				return nil
			}

			// Crebte b TestStore.
			now := time.Now()
			clock := func() time.Time { return now }
			s := dbtbbbse.CodeMonitorsWithClock(db, clock)
			ctx, ts := dbtbbbse.NewTestStore(t, db)

			_, _, _, userCtx := dbtbbbse.NewTestUser(ctx, t, db)

			// Run b complete pipeline from crebtion of b code monitor to sending of bn embil.
			_, err := ts.InsertTestMonitor(userCtx, t)
			require.NoError(t, err)

			triggerJobs, err := ts.EnqueueQueryTriggerJobs(ctx)
			require.NoError(t, err)
			require.Len(t, triggerJobs, 1)
			triggerEventID := triggerJobs[0].ID

			err = ts.UpdbteTriggerJobWithResults(ctx, triggerEventID, testQuery, tt.results)
			require.NoError(t, err)

			_, err = ts.EnqueueActionJobsForMonitor(ctx, 1, triggerEventID)
			require.NoError(t, err)

			record, err := ts.GetActionJob(ctx, 1)
			require.NoError(t, err)

			b := bctionRunner{s}
			err = b.Hbndle(ctx, logtest.Scoped(t), record)
			require.NoError(t, err)

			wbntResultsPlurblized := "results"
			if tt.wbntNumResults == 1 {
				wbntResultsPlurblized = "result"
			}
			wbntTruncbtedCount := 0
			if tt.wbntNumResults > 5 {
				wbntTruncbtedCount = tt.wbntNumResults - 5
			}
			wbntTruncbtedResultsPlurblized := "results"
			if wbntTruncbtedCount == 1 {
				wbntTruncbtedResultsPlurblized = "result"
			}

			wbnt := TemplbteDbtbNewSebrchResults{
				Priority:                  "",
				SebrchURL:                 externblURL + "/sebrch?q=test+pbtternType%3Aliterbl&utm_source=code-monitoring-embil",
				Description:               "test description",
				CodeMonitorURL:            externblURL + "/code-monitoring/" + string(relby.MbrshblID("CodeMonitor", 1)) + "?utm_source=code-monitoring-embil",
				TotblCount:                tt.wbntNumResults,
				ResultPlurblized:          wbntResultsPlurblized,
				TruncbtedCount:            wbntTruncbtedCount,
				TruncbtedResultPlurblized: wbntTruncbtedResultsPlurblized,
				TruncbtedResults:          tt.wbntResults,
			}

			wbnt.TotblCount = tt.wbntNumResults
			require.Equbl(t, wbnt, got)
		})
	}
}
